+++ src/OpenTelemetry.Resources.OperatingSystem/OperatingSystemCategory.cs
@@ -1,0 +1,13 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+namespace OpenTelemetry.Resources.OperatingSystem;
+
+internal enum OperatingSystemCategory
+{
+    Windows,
+    Android,
+    AppleOS,
+    Linux,
+    MacOS,
+}